Blank lines are included in diffs, fixed word handling
Including blank lines will hopefully minimize false negatives when some
lines didn't change, but only because they were the same as the last
update, when new text scrolled onto the page.
The code to check if a cell is in a word didn't take into account cells
with no content.