Change how streamed text is auto read
Instead of parsing it ourself, we send it to a blank screen, and let
vt100 handle it. The drawback of this is that, because tmux redraws the
screen instead of letting it scroll up, we can't read several screens of
text that quickly scrolled by. The upside is it eliminates all sorts of
parsing bugs.